Move containing_scope and ret_address to .cpp
